requirements.txt
altair==5.5.0
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.47
idna==3.13
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.20.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.7.0
protobuf==6.33.6
pyarrow==21.0.0
pydeck==0.9.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
